+++ src/components/select/index.ts
@@ -5,4 +5,8 @@
  * The copyright notice above does not evidence any
  * actual or intended publication of such source code.
  */
-export {Select, Props as SelectProps} from './select';
+export {Select, Props as SelectProps, SelectValueType} from './select';
+export {OptionLabel, Props as OptionLabelProps} from './option-label';
+export {Emitter, Props as EmitterProps} from './emitter';
+export {Option, Props as OptionProps} from './option';
+export {useSelectContext} from './select-context';