+++ src/components/select/select-native.module.css.d.ts
@@ -6,8 +6,6 @@
  * actual or intended publication of such source code.
  */
 
+export const root: string;
+
 export const select: string;
-
-export const hasError: string;
-
-export const emptyValue: string;