+++ src/components/select/select.tsx
@@ -1,120 +1,71 @@
 /*
- * Copyright (c) 2011-2018 Genestack Limited
+ * Copyright (c) 2011-2020 Genestack Limited
  * All Rights Reserved
  * THIS IS UNPUBLISHED PROPRIETARY SOURCE CODE OF GENESTACK LIMITED
  * The copyright notice above does not evidence any
  * actual or intended publication of such source code.
  */
-import classNames from 'classnames';
-import React, {PureComponent} from 'react';
 
-import * as styles from './select.module.css';
+import React from 'react';
 
-const emptyValue = '';
+import {MenuItem} from '../menu';
 
-type TargetProps = React.SelectHTMLAttributes<HTMLSelectElement>;
+import {Props as EmitterProps} from './emitter';
+import {SelectContext} from './select-context';
+import {SelectMenu} from './select-menu';
+import {SelectNative} from './select-native';
 
-interface Option {
-    value: any;
-    label: string;
+type TargetProps = React.SelectHTMLAttributes<HTMLSelectElement | HTMLDivElement>;
+
+/** Select value type */
+export type SelectValueType = string | number;
+
+/** Select change handler */
+export type OnChangeSelectHandler = (
+    value: SelectValueType,
+    event: React.ChangeEvent<HTMLSelectElement> | React.SyntheticEvent,
+    ref?: MenuItem
+) => void;
+
+/** All props for select */
+interface SelectWrapperProps {
+    /** Use native select instead Menu (default false) */
+    native?: boolean;
+    /** Select value */
+    value: SelectValueType;
+    /**
+     * Select onChange handler
+     * if native select is used, has arguments - value: string, React.ChangeEvent<HTMLSelectElement>
+     * if Menu is used, has arguments - value: string | number, React.SyntheticEvent, ref: MenuItem
+     */
+    onValueChange: OnChangeSelectHandler;
+    /**
+     * Select Emitter props
+     */
+    emitterProps?: Partial<EmitterProps>;
+    /** Sets `invalid` styles for select */
+    invalid?: boolean;
 }
 
 /** Select public properties */
-export interface Props extends TargetProps {
-    options: Option[];
-    value: any;
-    placeholder?: string;
-    hasError?: boolean;
-    onValueChange: (value: any) => any;
-    selectRef?: React.Ref<HTMLSelectElement>;
-}
+export type Props = SelectWrapperProps;
 
-/**
- * Select wrapper
- */
-export class Select extends PureComponent<Props> {
-    public static defaultProps: Partial<Props> = {
-        placeholder: 'Select a value...',
-        hasError: false
-    };
+/** Select wrapper */
+export function Select({native, emitterProps = {}, ...restProps}: Props & TargetProps) {
+    const Component = native ? SelectNative : SelectMenu;
+    const {invalid, intent, ghost} = emitterProps;
 
-    private handleChange: React.ChangeEventHandler<HTMLSelectElement> = (event) => {
-        const {onChange, onValueChange, options} = this.props;
-        const option = options[parseInt(event.target.value, 10)];
-
-        if (onChange) {
-            // original event has option index as its value,
-            // however in standard event it would be a stringified option.value
-            const eventValue = option ? String(option.value) : emptyValue;
-            onChange({
-                ...event,
-                target: {
-                    ...event.target,
-                    value: eventValue
-                }
-            });
-        }
-
-        if (onValueChange) {
-            const value = option ? option.value : emptyValue;
-            onValueChange(value);
-        }
-    };
-
-    public render() {
-        // "onChange" and "onValueChange" extracted here
-        // just to not allow them to be in the "rest" variable.
-        // "rest" is used to pass props down to native select-element
-        const {
-            value,
-            placeholder,
-            hasError,
-            required,
-            className,
-            options,
-            onChange,
-            onValueChange,
-            selectRef,
-            ...rest
-        } = this.props;
-
-        const isEmptyValue = value === emptyValue || value === undefined || value === null;
-
-        const selectHasError = hasError || (required && isEmptyValue);
-
-        const selectValue = isEmptyValue
-            ? emptyValue
-            : options.findIndex((option) => option.value === value);
-
-        const selectClassName = classNames(className, styles.select, {
-            [styles.emptyValue]: isEmptyValue,
-            [styles.hasError]: selectHasError
-        });
-
-        return (
-            <select
-                data-qa="select"
-                className={selectClassName}
-                value={selectValue}
-                onChange={this.handleChange}
-                required={required}
-                ref={selectRef}
-                {...rest}
-            >
-                {required && !isEmptyValue ? null : (
-                    <option value={emptyValue}>{placeholder}</option>
-                )}
-                {options.map((option, index) => {
-                    // we use index as a value to be able
-                    // to use actual option.value in "onValueChange"
-                    // (not its stringified copy)
-                    return (
-                        <option value={index} key={option.label}>
-                            {option.label}
-                        </option>
-                    );
-                })}
-            </select>
-        );
-    }
+    return (
+        <SelectContext.Provider
+            value={{
+                native,
+                ghost,
+                intent,
+                invalid,
+                disabled: restProps.disabled
+            }}
+        >
+            <Component emitterProps={emitterProps} {...restProps} />
+        </SelectContext.Provider>
+    );
 }