+++ src/utils/use-constol-invalidity.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2011-2019 Genestack Limited
+ * Copyright (c) 2011-2020 Genestack Limited
  * All Rights Reserved
  * THIS IS UNPUBLISHED PROPRIETARY SOURCE CODE OF GENESTACK LIMITED
  * The copyright notice above does not evidence any
@@ -7,18 +7,18 @@
  */
 import * as React from 'react';
 
-/** Subscribe to native input invalidity */
-export function useInputInvalidity(
-    inputRef: React.RefObject<HTMLInputElement>,
-    invalid: boolean | undefined
+/** Subscribe to native control invalidity */
+export function useControlInvalidity(
+    controlRef: React.RefObject<HTMLInputElement | HTMLSelectElement>,
+    invalid?: boolean
 ) {
-    // Input is valid by default
+    // Control is valid by default
     const [invalidState, setInvalidState] = React.useState(invalid || false);
 
-    // Use native input validation before browser paint
+    // Use native control validation before browser paint
     React.useLayoutEffect(() => {
-        if (inputRef.current) {
-            const nextInvalidState = !inputRef.current.validity.valid;
+        if (controlRef.current) {
+            const nextInvalidState = !controlRef.current.validity.valid;
             if (nextInvalidState !== invalidState) {
                 setInvalidState(nextInvalidState);
             }